Got the pop out windows working!
There is now a pop out with your results - a graph and customized text. The next to-do is to correct the API ping to try again if contacting fails, and also pop up an error if it keeps failing. Then, I also want to add to the results for a person with a pay cut an option to generate an email expressing their discontent.