Entry fields rewritten
The goal here is to do away with the hard coded entry fields, namely the last one that currently causes tons of complications. This is also standing in the way of adding rows at the END of the list instead of in the middle. The middle made sense when the end date was hard coded, but now it is unnecessary after making end dates flexible. Now that there is a load/save function the best application of it would be to load and update your pay info each year. The "insert between" doesn't allow that.

The new system will take advantage of the same function as the "row add" button that was created to build a dictionary of entry fields whilst also storing the data in a simpler system than before.

Upon this commit I've got the entry field system redone. However, the validator system still needs overhaul.